[nrf fromlist] cmake: multi image build support
Add support for non-recursive single-toolchain multi-image
builds. This resolves #7868.

We are seeing the need to boot Zephyr applications in multiple
stages, a real-world use-case of even a 3-stage bootloader has
been identified and tested.

Each bootloader needs to be individually upgrade-able and have
it's own configurations of Zephyr libraries. To achieve this each
bootloader is organized as it's own executable.

The problem is that the build system can only build one
executable. This creates a usability issue as the user must invoke a
large set of arcane commands to build, sign, and flash each
bootloader.

To resolve this we re-organize the build system such that it can build
multiple executables.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>